Yes or no?

+++ scripts/markers/unmapped_markers.pl
@@ -1,7 +1,9 @@
-#! usr/ bin/perl
+#! usr/local/ensembl/bin/perl
+
 use strict;
 use warnings;
 $| = 1;
+
 use Getopt::Long;
 use Bio::EnsEMBL::UnmappedObject;
 use Bio::EnsEMBL::DBSQL::UnmappedObjectAdaptor;
@@ -17,6 +19,8 @@ my @duplicates;
 my @unmapped;
 my $pass = '';
 my $ln = 'Marker';
+my $dry_run = 0;
+
 GetOptions(
 	   'host|dbhost:s'      => \$host,
 	   'user|dbuser:s'      => \$user,
@@ -25,6 +29,7 @@ GetOptions(
 	   'logic_name:s'       => \$ln,
 	   'pass|dbpass:s'      => \$pass,
 	   'max_duplicates:n'   => \$max_duplicates,
+           'dry_run!'           => \$dry_run,
 	  );
 
 unless ($host && $user && $dbname && $port){
@@ -50,11 +55,18 @@ my $db = new Bio::EnsEMBL::DBSQL::DBAdaptor
 
 die("Cannot open connection to $dbname:$host.$port for  $user\n") unless $db;
 
+if ($dry_run) {
+  print STDERR "This is a dry run.\n";
+}
+
+print "You will be removing markers which map to the genome more than $max_duplicates times.\n";
+
 my $aa = $db->get_AnalysisAdaptor;
 my $analysis = $aa->fetch_by_logic_name($ln);
 die("Cannnot find analysis with logic name $ln \n") unless $analysis;
 my $ma = $db->get_MarkerAdaptor;
 my $umma = $db->get_UnmappedObjectAdaptor;
+
 # fetch Markers
 my @markers = @{$ma->fetch_all};
 die("No markers found\n") unless  scalar(@markers) > 0 ;
@@ -62,11 +74,12 @@ print STDERR "Found " . scalar(@markers) . " markers, checking / setting marker_
 foreach my $marker ( @markers ) {
   my @features = @{$marker->get_all_MarkerFeatures} ;
   my @map_locations = @{$marker->get_all_MapLocations} ;
-  if ( scalar(@features > $max_duplicates ) && scalar(@map_locations) ==  0 ) {
+  if ( (scalar@features > $max_duplicates) && scalar(@map_locations) ==  0 ) {
     push @duplicates,$marker;
   } elsif ( scalar(@features == 0 ) && scalar(@map_locations) ==  0 ) {
     push @unmapped,$marker;
   }
+
   # update map weights
   foreach my $feat (@features) {
     if ( scalar(@features) != $feat->map_weight ) {
@@ -77,38 +90,44 @@ foreach my $marker ( @markers ) {
 
 print STDERR "Found " . scalar(@duplicates) . " duplicates and " . scalar(@unmapped) . " unmapped markers\n";
 
-print STDERR "Creating unmapped entries...\n";
-
-my @unmapped_objects = @{make_entries("duplicates",\@duplicates)};
-push @unmapped_objects,@{make_entries("unmapped",\@unmapped)};
-foreach my $umo ( @unmapped_objects ) {
-  $umma->store($umo);
-}
-print STDERR "Deleting duplicated and unmapped markers...\n";
-
-# delete the markers that dont map 
-foreach my $marker ( @duplicates ) {
-  sql("DELETE from marker where marker_id = " . $marker->dbID . ";",$db);
-}
-foreach my $marker ( @unmapped ) {
-  sql("DELETE from marker where marker_id = " . $marker->dbID . ";",$db);
+if ($dry_run) {
+  print STDERR "Dry run finished. Unmapped entries not created.  No markers or marker_features have been deleted from the DB.\n";
 }
 
-# tidy up associated tables
-# marker feature
-sql("DELETE marker_feature from marker_feature LEFT JOIN marker
-ON marker_feature.marker_id =  marker.marker_id 
-WHERE marker.marker_id is null ",$db);
-# marker feature where the marker hits lots of times but is also in the marker_map_locations table
-# just delete the multiple hitting markers
-sql("DELETE marker_feature from marker_feature 
-WHERE map_weight > $max_duplicates",$db);
-# marker synonym
-sql("DELETE marker_synonym from marker_synonym LEFT JOIN marker
-ON marker_synonym.marker_id =  marker.marker_id 
-WHERE marker.marker_id is null",$db);
+if (!$dry_run) {
+  print STDERR "Creating unmapped entries...\n";
 
-print STDERR "Done\n";
+  my @unmapped_objects = @{make_entries("duplicates",\@duplicates)};
+  push @unmapped_objects,@{make_entries("unmapped",\@unmapped)};
+  foreach my $umo ( @unmapped_objects ) {
+    $umma->store($umo);
+  }
+  print STDERR "Deleting duplicated and unmapped markers...\n";
+
+  # delete the markers that dont map 
+  foreach my $marker ( @duplicates ) {
+    sql("DELETE from marker where marker_id = " . $marker->dbID . ";",$db);
+  }
+  foreach my $marker ( @unmapped ) {
+    sql("DELETE from marker where marker_id = " . $marker->dbID . ";",$db);
+  }
+
+  # tidy up associated tables
+  # marker feature
+  sql("DELETE marker_feature from marker_feature LEFT JOIN marker
+  ON marker_feature.marker_id =  marker.marker_id 
+  WHERE marker.marker_id is null ",$db);
+  # marker feature where the marker hits lots of times but is also in the marker_map_locations table
+  # just delete the multiple hitting markers
+  sql("DELETE marker_feature from marker_feature 
+  WHERE map_weight > $max_duplicates",$db);
+  # marker synonym
+  sql("DELETE marker_synonym from marker_synonym LEFT JOIN marker
+  ON marker_synonym.marker_id =  marker.marker_id 
+  WHERE marker.marker_id is null",$db);
+
+  print STDERR "Done\n";
+}
 
 exit;
